[build] set VSIX compression level to Normal
While generating a VSIX, I noticed this in the build log:

    CreateVsixContainer:
      Creating VSIX Container...
      Creating the package with the following compression level. "NotCompressed".

Changing the `ZipPackageCompressionLevel` property to `Normal`, seems
to change the VSIX so it is properly compressed.

The file size goes from 584991368 bytes to 172998318 bytes.

It's worth doing this, just to make the download smaller in the OSS
repo. I need to do some further research and make sure the VSIX we are
shipping commercially is compressed.