[Xamarin.Android.Build.Tasks] refactor/split up ResolveSdks (#1916)
Context: http://work.devdiv.io/639234

The `ResolveSdks` MSBuild task was quite large, and had lots of inputs
and outputs. Because of this, it is difficult to tell which parts of
it are slow.

As a first stab at refactoring, I split it into three tasks that are
more focused in their concerns:
- `ResolveSdks` - now focuses on resolving `AndroidSdkPath`,
  `AndroidNdkPath` and `JavaSdkPath`. It also sets up any static
  members of `MonoAndroidHelper`.
- `ValidateJavaVersion` - shells out to `java` and `javac`, sets
  `MinimumRequiredJdkVersion` and `JdkVersion`.
- `ResolveAndroidTooling` - does all the other work: calculates
  `TargetFrameworkVersion` and finds the paths of various Android
  command line tooling

The goal here is to clean things up without really changing how
anything works. And in `ResolveSdksTaskTests`, I kept the tests
exactly the same but changed it to call all three tasks.

After the refactoring, we have a much better picture of what we need
to improve performance for. Here is the result of running the simplest
test, `BuildBasicApplication`:

    ResolveSdks 21ms
    ValidateJavaVersion 298ms
    ResolveAndroidTooling 15ms

So `ValidateJavaVersion` is where we should focus on performance
improvements (caching).

~~ Improvements to logging ~~

In general we have been logging like this for every MSBuild task:

    public override bool Execute ()
    {
        Log.LogDebugMessage ($"{nameof (MyMSBuildTask)}");
        Log.LogDebugMessage ($"  {nameof (InputValue1)}: {InputValue1}");

        //Do stuff

        Log.LogDebugMessage ($"{nameof (MyMSBuildTask)} Outputs:");
        Log.LogDebugMessage ($"  {nameof (OutputValue1)}: {OutputValue1}");

        return !Log.HasLoggedErrors;
    }

The diagnostic log output would look something like:

    Task "MyMSBuildTask" (TaskId:8)
      Task Parameter:InputValue1=1234 (TaskId:8)
      MyMSBuildTask: (TaskId:8)
        InputValue1: 1234 (TaskId:8)
      MyMSBuildTask Outputs: (TaskId:8)
        OutputValue1: abcd (TaskId:8)
      Output Property: OutputValue1=abcd (TaskId:8)
    Done executing task "MyMSBuildTask". (TaskId:8)

The log messages are mostly duplicate, and seem unneeded. However, the
`Output Property` log only prints from MSBuild if this is specified in
the target:

    <Output TaskParameter="OutputValue1" PropertyName="OutputValue1" />

Since you *could* omit this, or rename the property, we should
continue logging all `[Output]` properties. However, I have removed
input ones in the changed tasks since they seem completely duplicate.

Other changes:
- Removed `$(_TargetFrameworkVersion)` property from
  `Xamarin.Android.Common.targets` in favor of just using
  `$(TargetFrameworkVersion)`. This seemed to be leftover from the
  past and not used.
- General refactoring/ordering of properties on MSBuild tasks
- General refactoring/ordering of properties in
  `Xamarin.Android.Common.targets` / `Xamarin.Android.Bindings.targets`
- Use `nameof` where appropriate in log messages
- `ResolveAndroidTooling` should use `Path.PathSeparator`
- Updated `ValidateUseLatestAndroid` and `CheckTargetFrameworkVersion`
  tests