[timing] use Release build of xabuild/xamarin-android if possible
When timing builds, and plotting them on Jenkins, we should prefer to
use the `Release` build output of `xamarin-android` if it exists.

This also will fix a problem downstream in `monodroid`, where I'm not
able to profile `Debug` builds because `bin\Debug` *might* not exist.
`monodroid` seems to mostly use `Release` builds of `xamarin-android`.
We really need to profile `Debug` Xamarin.Android application builds
so that we can see how `Fast Deployment` is performing.