Ignore <merge> when generating code-behind source (#1937)
Fixes: https://github.com/xamarin/xamarin-android/issues/1929

`<merge>` is an element used by Android to avoid stacking too many view groups
when one layout is included into another. In such cases `<merge>` can be used as
a place-holder which is replaced by the enclosing view group of the layout that
includes another layout with `<merge>` as the root element. However, as much as
you can call `FindViewById` to find an instance of `<merge>` which was given the
`android:id` attribute, Android will always return `null` in such case, as the
element has no type counterpart in the `Android.Views` collection of views.
Therefore, we must ignore the element whenever we process a layout that contains
it decorated with `android:id` in order to avoid confusion in generated
code (i.e. the property generated for the element would be of type `merge` which
does not exist). We could instead generate a property of type
`Android.Views.View` and always return `null` from it, but that unnecessarily
increases code size - we'd be adding effectively dead code.