Include soft debugger interface code in libmonodroid
Due to a mistake in `CMakeLists.txt` the `DEBUG` macro wasn't defined and, in
effect, no libmonodroid.so built after the switch to cmake included the soft
debugger code required for debugging on device or in the emulator.

This commit fixes the issue by properly defining `-DDEBUG` whenever necessary.
Additionally, it now properly adds other native debugging flags which were
previously mistakenly omitted.