Minimal changes to fix libmonodroid build for Windows with DEBUG
When the `DEBUG` macro is defined it causes the build to include code that's
specific to Unix systems and will not build when targetting Windows (mostly
related to sockets). This commit modifies the minimal set of lines (on purpose,
to reduce the noise) to make the Windows cross-builds work.

Additionally, the `--cache-file` configure parameter is removed from
`mono-runtimes.projitems` as it is a duplicate and incorrect, with the correct
one found in `mono-runtimes.targets` file, in the `_ConfigureCrossRuntimes`
target.