Remove local reference to a Java object when done with it
Fixes: https://devdiv.visualstudio.com/DevDiv/_workitems/edit/627319

Network interface information retrieval code needs to call a number of Java
methods in order to obtain the information from the OS. A failure to free one of
the retrieved local references caused a leak and, in effect, DevDiv issue

Make sure to remove the refernce before returning to the caller.

Lumped together is a potential NULL dereference fix and an unnecessary signed
comparison of unsigned operands.