[Xamarin.Android.Build.Tasks] ManagedResourceParser should not throw
Fixes https://devdiv.visualstudio.com/DevDiv/Xamarin%20Project%20Mgrs/_workitems/edit/639674

If a user generates an invalid or incomplete xml file the
`ManagedResourceParser` should not crash with an exception.
It should issue a warning and let the build fail at a point
where `aapt` can issue a decent error.

If we leave the error reporting to `aapt` users should know how
to fix the issue, and get a decent error message.