[Xamarin.Android.Build.Tasks] Set the timestamps on the `resources.cache` files.
We currently do not force the setting of the write and access
timestamps when saving the resource.caches files. We have
recently (90e31550) been reviewing and adding this in other
parts of the build system. This is to ensure that when we do
modify a file it has the latest timestamp.

This commit adds support for calling

	MonoAndroidHelper.SetLastAccessAndWriteTimeUtc ()

when we save the XDocument cache files.