[msbuild/tests] add test for custom framework setup via NuGet package.
It proves incomplete but kind of working solution for https://github.com/xamarin/xamarin-android/issues/1880 .

The original idea (and what once worked) was to only extend an MSBuild
property for ResolveSdks task. But after a couple of changes it does not
work anymore, as there had been a handful of SDK resolution changes.

Since it's not really safe to depend on those properties, we should rather
provide a complete set of framework assemblies (which means, packages
containing both v1.0 and latest) to get app still build.