[Xamarin.Android.Build.Tasks] Fix issues with `ExtraArgs` for Aapt2.
Fixes #1985

We currently use the same `AndroidResgenExtraArgs` property for
both `aapt` and `aapt2`. As a result when passing `--no-crunch`
to the `aapt2 link` command the following error is raised.

	error APT0000: unknown option '--no-crunch'. "unknown option '--no-crunch'.".

Even worse.. you get 70 errors as `aapt2` dumps the contents of
the help output to stdout. This is not idea and does not allow
the user to act on the error.

So this commit does a couple of things

1) Add two new properties for passing arguments for `aapt2`

	<AndroidAapt2LinkExtraArgs/>
	<AndroidAapt2CompileExtraArgs/>

   These will be used for `aapt2` instead of `AndroidResgenExtraArgs`.

2) Update the `Aapt2` to handle the unknown option error by emiting
   and more descriptive error message. Also make sure we ignore the
   rest of the `help` messages so that the user gets only one error
   and not 70.