[Mono.Android] Refactor dynamic method names creation
Put the dynamic method name logic in one place. It also removes the
formatted string use and uses `InvariantCulture`.

It hopefully improves the performance (simpler string operations,
InvariantCulture). The change in performance is so small though, that
I wasn't able to measure it.

We use `long` counter, because `ConstructorBuilder` can, in some
cases, create a new constructor per java object instance. This should
be addressed in the future. After that we can probably get back to
`int` counter.