[Xamarin.Android.Build.Tasks] Fix Aapt2 warnings showing as errors
Fixes (again) #1770

On windows we are hitting the same issue where the warnings
`aapt2` is producing are being logged as errors.

The weird part is that the "file" group was not being populated
with the right informaton.. This turns out to be down to how
`aapt2` is reporting the warnings..

	W/ResourceType(23681): For resource 0x0101053d, entry index(1341) is beyond type entryCount(733)
	W/ResourceType( 5536): For resource 0x0101053d, entry index(1341) is beyond type entryCount(1155)
		       ^
		    Hmmmmm.

This space throws off our regex and results in the entire line
ending up in the `message` property.
So the fix is to update the regex to allow spaces for the `line`
capture group. We then also need to Trim that value before passing
it into `int.Parse` to ensure we don't get an error.