[monodroid] Recreate shared runtime symlink when invalid (#1984)
Background: Within Commercial Xamarin.Android, Fast Deployment works
by having the [shared runtime provider][0] pass the path to the
`Mono.Android.DebugRuntime` Android package into `libmonodroid.so`.
`libmonodroid.so` will later
[create a symlink to the `libmonosgen-2.0.so`][1] "hosted" by the
`Mono.Android.DebugRuntime` package.  This allows Debug `app.apk`
files to be smaller, as they don't need to contain
`libmonosgen-2.0.so`; it is instead provided by the shared runtime.

Additionally, on recent Android versions the installation directory
of a `.apk` file contains an embedded "uniquifier":

	$ adb shell pm path Mono.Android.DebugRuntime
	package:/data/app/Mono.Android.DebugRuntime-ox9_8IlcD_-F05cZhNjAPA==/base.apk

Here, `-ox9_8IlcD_-F05cZhNjAPA==` is the uniquifier.

The uniquifier value will change every time whenever the associated
package is updated or reinstalled.

These two facts interact in an unfortunate manner:

 1. Build and run a `Debug` mode Xamarin.Android app
 2. Update Xamarin.Android, e.g. from Xamarin.Android 8.3 to 9.0.
 3. Build and run again

On the second deployment (3), a *new* shared Mono runtime will be
installed on the device.  This is *expected* because the shared Mono
runtime contains e.g. `mscorlib.dll` and other assemblies which were
changed by (2).

Unfortunately, the uniquifier for the `Mono.Android.DebugRuntime`
package *will also change*, meaning that the symlink that
`libmonodroid.so` created in step (1) is now *invalid*.

	$ run-as com.xamarin.myapp ls -la /data/user/0/com.xamarin.myapp/files/.__override__/links
	lrwxrwxrwx 1 u0_a132 u0_a132   93 2018-07-19 16:57 libmonosgen-2.0.so -> /data/app/Mono.Android.DebugRuntime-Y2eHYda6s0hMKJmiCfcYlA==/lib/x86/libmonosgen-32bit-2.0.so

	$ ls /data/app/Mono.Android.DebugRuntime-Y2eHYda6s0hMKJmiCfcYlA==/lib/x86/
	ls: /data/app/Mono.Android.DebugRuntime-Y2eHYda6s0hMKJmiCfcYlA==/lib/x86/: No such file or directory

This causes the app to crash on startup:

	[monodroid] Creating public update directory: `/data/user/0/com.xamarin.myapp/files/.__override__`
	[monodroid] Using override path: /data/user/0/com.xamarin.myapp/files/.__override__
	[monodroid] Trying to load sgen from: /data/user/0/com.xamarin.myapp/files/.__override__/libmonosgen-2.0.so
	[monodroid] Trying to load sgen from: /storage/emulated/0/Android/data/com.xamarin.myapp/files/.__override__/libmonosgen-2.0.so
	[monodroid] Trying to load sgen from: /storage/emulated/0/../legacy/Android/data/com.xamarin.myapp/files/.__override__/libmonosgen-2.0.so
	[monodroid] Trying to load sgen from: /data/app/com.xamarin.myapp-5PqDQ9Z3Et8G7_1D2EtGIg==/lib/x86/libmonosgen-2.0.so
	[monodroid] Trying to load sgen from: /data/user/0/com.xamarin.myapp/files/.__override__/links/libmonosgen-2.0.so
	[monodroid] Trying to load sgen from: /system/lib/libmonosgen-2.0.so
	[monodroid] cannot find libmonosgen-2.0.so in override_dir: /data/user/0/com.xamarin.myapp/files/.__override__, app_libdir: /data/app/com.xamarin.myapp-5PqDQ9Z3Et8G7_1D2EtGIg==/lib/x86 nor in previously printed locations.
	[monodroid] Do you have a shared runtime build of your app with AndroidManifest.xml android:minSdkVersion < 10 while running on a 64-bit Android 5.0 target? This combination is not supported.
	[monodroid] Please either set android:minSdkVersion >= 10 or use a build without the shared runtime (like default Release configuration).

The fix is to update `libmonodroid.so` so that if the **symlink**(2)
call *fails* with `EEXIST` -- meaning the symlink file already exists
but, because `file_exists(link)` failed, we deduce the the
symlink target no longer exists --  we delete the symlink and try to
create the symlink again.

[0]: https://github.com/xamarin/java.interop/blob/8298c7222b9e6ea13b821ddc117e4dae4baf7d2f/src/Java.Interop.Tools.JavaCallableWrappers/Resources/MonoRuntimeProvider.Shared.java
[1]: https://github.com/xamarin/xamarin-android/blob/9be682a448e166db359eb99be16143fa23c73123/src/monodroid/jni/monodroid-glue.c#L656-L669