[Xamarin.Android.Build.Tasks] Fix Aapt2 warnings showing as errors (#2027)
Fixes (again): https://github.com/xamarin/xamarin-android/issues/1770

On Windows we are hitting the same issue where the warnings `aapt2`
is producing are being logged as errors.

The weird part is that the "file" group was not being populated
with the right information.  This turns out to be down to how
`aapt2` is reporting the warnings: there may be *spaces* within the
line number portion:

	W/ResourceType(23681): For resource 0x0101053d, entry index(1341) is beyond type entryCount(733)
	W/ResourceType( 5536): For resource 0x0101053d, entry index(1341) is beyond type entryCount(1155)
		       ^

This space throws off our regex and results in the entire line ending
up in the `message` property.  The fix is to update the regex to
allow spaces for the `line` capture group.  We then also need to Trim
that value before passing it into `int.Parse()` to ensure we don't
get an error.