[Xamarin.Android.Build.Tasks] CopyIfChanged sets LastWriteTimeUtc (#2038)
`MonoAndroidHelper.CopyIfChanged()` has a boolean return value,
returning true if the destination file was updated, otherwise false.

Question: What should be done with this return value?

This is, surprisingly, a difficult question to answer: sometimes the
return value is checked, more often it isn't.  When it is checked,
it's to determine whether or not
`MonoAndroidHelper.SetLastAccessAndWriteTimeUtc()` should be called;
see e.g. commit d7e3a232.

So should `SetLastAccessAndWriteTimeUtc()` *always* be invoked?
Sometimes?  Never?  Is there a guideline *at all*?

Turns out...there wasn't a guideline that we could infer.
Additionally, the *lack* of `SetLastAccessAndWriteTimeUtc()` was a
*source* of bugs, as `File.Copy()` would *preserve* the timestamp
of the *source* file, while many targets want "updated" timestamps
to prevent the target from re-running again; again, d7e3a232.

This entire scenario is confusing.  Confusion is *not* good.

To try to clear up this confusion, a question:  Is there any scenario
for which `MonoAndroidHelper.CopyIfChanged()` *shouldn't* call
`MonoAndroidHelper.SetLastAccessAndWriteTimeUtc()` when the source
file needs to be copied to the target file?

We cannot currently imagine such a scenario.

Alter the semantics of `MonoAndroidHelper.CopyIfChanged()` so that
the timestamp of the target file is updated to "now" when the target
file is updated (i.e. the source file differs).

This *removes* many of our "update the write time if
`CopyIfChanged()` returned true" codepaths, and makes it considerably
easier to reason about what *exactly* should be done with the return
value of `MonoAndroidHelper.CopyIfChanged()`: it is now usually safe
to ignore, but if e.g. two files are "linked" in some way, it can
continue to be used.