[Xamarin.Android.Build.Tasks] fix for parallel builds
Context: https://github.com/xamarin/xamarin-android/issues/2043

When parallel builds occur, we can get `obj/Debug/lp` into a broken
state:

    Resources/layout/abc_action_menu_layout.xml warning XA1001:
        AndroidResgen: Warning while updating Resource XML '/var/folders/nx/rzr6lcqj6j3_vp7vtp28v2cw0000gn/T/tmp6bafd8e4.tmp': Root element is missing.

Then later, `<Aapt />` fails on the empty file:

    obj/Debug/lp/11/jl/res/layout/abc_action_menu_layout.xml(2,0): error APT0000: Error parsing XML: no element found

It appears that `ConvertResourcesCases` still copies the file, even if
`AndroidResource.UpdateXmlResource` fails. Subsequent builds will
continue to fail, because an empty file has been copied with a newer
timestamp. Once in this state, developers would have to `Rebuild` the
project or delete `obj`.

The fix here is to:
- Return a `bool` indicating if `AndroidResource.UpdateXmlResource`
  succeeds
- Don't write the final file if `AndroidResource.UpdateXmlResource`
  failed

This change doesn't fully fix #2043, as I think VS for Mac should
prevent builds from happening in parallel. I tried the change in the
IDE once, and it worked. However, I feel like if I repeated the steps
10 times in a row, something would eventually go wrong.

This change *does*, however, fix the issue for the user so a `Rebuild`
is not required.

Other changes:
- The original temp file can use `File.Copy` since it is empty
- We can  use the `?` operator for invoking `logMessage`